Rename utils.py to channel_positioning.py and update all imports
