+++ pylabrobot/liquid_handling/backends/hamilton/STAR_backend.py
@@ -61,7 +61,7 @@ from pylabrobot.liquid_handling.standard import (
   SingleChannelAspiration,
   SingleChannelDispense,
 )
-from pylabrobot.liquid_handling.utils import (
+from pylabrobot.liquid_handling.channel_positioning import (
   MIN_SPACING_EDGE,
   get_tight_single_resource_liquid_op_offsets,
   get_wide_single_resource_liquid_op_offsets,
