+++ pylabrobot/resources/container_tests.py
@@ -1,6 +1,7 @@
 import json
 import unittest
 
+from pylabrobot.liquid_handling.errors import ChannelsDoNotFitError
 from pylabrobot.serializer import serialize
 
 from .container import Container
@@ -208,7 +209,7 @@ class TestNoGoZoneCollision(unittest.TestCase):
     return Container(name="c", size_x=10, size_y=size_y, size_z=10, no_go_zones=no_go_zones)
 
   def test_no_zones_uses_standard_spread(self):
-    from pylabrobot.liquid_handling.utils import compute_channel_offsets
+    from pylabrobot.liquid_handling.channel_positioning import compute_channel_offsets
 
     c = self._make_container(90)
     result = compute_channel_offsets(c, num_channels=1)
@@ -217,7 +218,7 @@ class TestNoGoZoneCollision(unittest.TestCase):
     self.assertAlmostEqual(result[0].y, 0.0)
 
   def test_1_channel_in_2_compartments(self):
-    from pylabrobot.liquid_handling.utils import compute_channel_offsets
+    from pylabrobot.liquid_handling.channel_positioning import compute_channel_offsets
 
     # 90mm container, divider at Y=44-46 -> 2 compartments [0,44] and [46,90]
     # edge_clearance = 2.0
@@ -234,7 +235,7 @@ class TestNoGoZoneCollision(unittest.TestCase):
     self.assertAlmostEqual(result[0].y, 23.0)
 
   def test_2_channels_across_2_compartments(self):
-    from pylabrobot.liquid_handling.utils import compute_channel_offsets
+    from pylabrobot.liquid_handling.channel_positioning import compute_channel_offsets
 
     c = self._make_container(
       90,
@@ -246,7 +247,7 @@ class TestNoGoZoneCollision(unittest.TestCase):
     self.assertGreater(result[0].y, result[1].y)
 
   def test_4_channels_across_2_compartments(self):
-    from pylabrobot.liquid_handling.utils import compute_channel_offsets
+    from pylabrobot.liquid_handling.channel_positioning import compute_channel_offsets
 
     c = self._make_container(
       90,
@@ -256,18 +257,18 @@ class TestNoGoZoneCollision(unittest.TestCase):
     self.assertEqual(len(result), 4)
 
   def test_raises_when_impossible(self):
-    from pylabrobot.liquid_handling.utils import compute_channel_offsets
+    from pylabrobot.liquid_handling.channel_positioning import compute_channel_offsets
 
     # Entire container is no-go
     c = self._make_container(
       12,
       no_go_zones=[(Coordinate(0, 0, 0), Coordinate(10, 12, 10))],
     )
-    with self.assertRaises(ValueError):
+    with self.assertRaises(ChannelsDoNotFitError):
       compute_channel_offsets(c, num_channels=1)
 
   def test_3_compartments_6_channels(self):
-    from pylabrobot.liquid_handling.utils import compute_channel_offsets
+    from pylabrobot.liquid_handling.channel_positioning import compute_channel_offsets
 
     # 150mm container, 2 dividers -> 3 compartments, 6 channels -> 2 per compartment
     c = self._make_container(
